Reorganize entrypoint
